core: refactored company balance event
